First mostly fuctional implementation This needs a lot of optimization...
