+++ .drone.yml
@@ -1,0 +1,10 @@
+build:
+  image: golang
+  commands:
+    - go get
+    - go build
+    - go test
+    - tar cfJ pgGlaskugel.tar.xz pgglaskugel
+    - GOOS=darwin GOARCH=amd64 go build -o pgglaskugel.app main.go
+    - tar cfJ pgGlaskugel.app.tar.xz pgglaskugel.app
+#    - GOOS=windows GOARCH=amd64 go build -o pgGlaskugel.exe main.go
