+++ .vscode/launch.json
@@ -1,0 +1,16 @@
+{
+    "version": "0.2.0",
+    "configurations": [{
+        "name": "Launch",
+        "type": "go",
+        "request": "launch",
+        "mode": "debug",
+        "remotePath": "",
+        "port": 2345,
+        "host": "127.0.0.1",
+        "program": "${fileDirname}",
+        "env": {},
+        "args": [],
+        "showLog": true
+    }]
+}
